+++ eras/shelley/impl/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.5.0.0
 
+* Introduce `ShelleyPoolreapEnv` and replace `Environment` for `ShelleyPOOLREAP`
 * Remove redundant reimplementation of lens function `%~`: `updateWithLens`.
 * Change `getConstituitionHash` to `getConstitution`
 * Replace `constitutionHash` with `constitutionAnchor`